+++ module-apps/application-settings-new/windows/InputLanguageWindow.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "BaseSettingsWindow.hpp"
+
+namespace gui
+{
+
+    class InputLanguageWindow : public BaseSettingsWindow
+    {
+      public:
+        InputLanguageWindow(app::Application *app);
+        void buildInterface() override;
+    };
+} // namespace gui