+++ module-apps/application-settings-new/windows/LockedScreenWindow.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "BaseSettingsWindow.hpp"
+
+namespace gui
+{
+    class LockedScreenWindow : public BaseSettingsWindow
+    {
+      public:
+        LockedScreenWindow(app::Application *app);
+        void buildInterface() override;
+    };
+} // namespace gui