+++ changelog.md
@@ -6,6 +6,7 @@
 
 * `[calendar]` Add notifications from database handling.
 * `[cellular]``[call app]` Added DTMF tone generation 
+* `[settings]` Add basic display and keypad options window
 * `[audio]` Enabled sounds priorities
 
 ### Changed
@@ -13,7 +14,6 @@
 
 
 ### Fixed
-
 
 
 ### Other